backend/requirements.txt
flask
flask-cors
watchdog
python-dotenv
openai
google-generativeai